[RLlib] Feature importance estimate handles single policy MultiAgentBatch.
