+++ rllib/offline/estimators/off_policy_estimator.py
@@ -4,11 +4,7 @@ import tree
 from typing import Dict, Any, List
 
 import logging
-from ray.rllib.policy.sample_batch import (
-    MultiAgentBatch,
-    DEFAULT_POLICY_ID,
-    SampleBatch,
-)
+from ray.rllib.policy.sample_batch import SampleBatch
 from ray.rllib.policy import Policy
 from ray.rllib.utils.policy import compute_log_likelihoods_from_input_dict
 from ray.rllib.utils.annotations import (
@@ -192,33 +188,6 @@ class OffPolicyEstimator(OfflineEvaluator):
         estimates["v_delta"] = estimates["v_target"] - estimates["v_behavior"]
 
         return estimates
-
-    @DeveloperAPI
-    def convert_ma_batch_to_sample_batch(self, batch: SampleBatchType) -> SampleBatch:
-        """Converts a MultiAgentBatch to a SampleBatch if neccessary.
-
-        Args:
-            batch: The SampleBatchType to convert.
-
-        Returns:
-            batch: the converted SampleBatch
-
-        Raises:
-            ValueError if the MultiAgentBatch has more than one policy_id
-            or if the policy_id is not `DEFAULT_POLICY_ID`
-        """
-        # TODO: Make this a util to sample_batch.py
-        if isinstance(batch, MultiAgentBatch):
-            policy_keys = batch.policy_batches.keys()
-            if len(policy_keys) == 1 and DEFAULT_POLICY_ID in policy_keys:
-                batch = batch.policy_batches[DEFAULT_POLICY_ID]
-            else:
-                raise ValueError(
-                    "Off-Policy Estimation is not implemented for "
-                    "multi-agent batches. You can set "
-                    "`off_policy_estimation_methods: {}` to resolve this."
-                )
-        return batch
 
     @DeveloperAPI
     def check_action_prob_in_batch(self, batch: SampleBatchType) -> None:
